new kdtree

+++ openpoints/dataset/s3dis/s3dis.py
@@ -151,7 +151,7 @@ class S3DIS(Dataset):
 
         if 'heights' not in data.keys():
             data['heights'] =  torch.from_numpy(coord[:, self.gravity_dim:self.gravity_dim+1].astype(np.float32))
-        data['x'] = torch.cat([data['x'], data['heights']], dim=-1).transpose(0, 1).contiguous()
+        data['x'] = torch.cat([data['x'], data['heights']], dim=-1).transpose(0, 1).contiguous().float()
         return self.make_idx(idx, data)
 
     def __len__(self):
@@ -159,7 +159,7 @@ class S3DIS(Dataset):
         # return 1   # debug
 
     def make_idx(self, idx, data):
-        coord = data['pos']
+        coord = data['pos'].float()
         gs = NaiveGaussian3D(self.gs_opts, batch_size=8, device=coord.device)
         gs.projects(coord, cam_seed=idx, cam_batch=gs.opt.n_cameras * 2)
         visible = gs.gs_points.visible.squeeze(1).float()
